third_party/blink/web_tests/platform/linux-chrome/external/wpt/webtransport/sendorder.https.any.sharedworker-expected.txt

This is a testharness.js-based test.
[FAIL] WebTransport client should be able to create and handle a bidirectional stream with sendOrder
  assert_equals: expected (number) 3 but got (undefined) undefined
[FAIL] WebTransport client should be able to modify unset sendOrder after stream creation
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] WebTransport client should be able to modify existing sendOrder after stream creation
  assert_equals: expected (number) 3 but got (undefined) undefined
[FAIL] WebTransport sendorder should not starve a stream without sendorder
  assert_equals: expected (number) 1 but got (undefined) undefined
[FAIL] WebTransport sendorder should starve a lower priority stream
  assert_equals: expected (number) 1 but got (undefined) undefined
[FAIL] WebTransport sendorder should starve a lower priority stream, variant 2
  assert_equals: expected (number) 1 but got (undefined) undefined
Harness: the test ran to completion.
